Update Communication presentation, WAYC scoreboard
Final version of the communication presentation.  Ended up not being
used.

Final version of the WAYC scoreboard.  This is the version that was used
in the S1-691-15 "Game of Life" presentation on Feb. 28, 2015.  It
worked surprisingly well. Only a few minor hiccups that I was able to
work around fairly easily.